Added note on Visual C++ Redistributable to README
The Windows wheels are built with the glfw shared libraries built for VC 2010 (32-bit) or VC 2012 (64-bit) and the redistributables for these are required to load the libraries. Most users already have those installed, but this commit adds a note to the README and links to the official downloads.

Fixes #23.